Moved DataPlotter into a separate file, moved drawing code to single data objects
